requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
astroid==2.13.2
autopep8==2.0.1
boto3==1.26.45
botocore==1.29.45
cachetools==5.2.0
certifi==2022.12.7
colorama==0.4.6
dill==0.3.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.11.4
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pycodestyle==2.10.0
pylint==2.15.9
python-dateutil==2.8.2
python-telegram-bot==20.0a2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
wrapt==1.14.1
